+++ mowl/projection/__init__.py
@@ -4,4 +4,5 @@ from .owl2vec_star.model import OWL2VecStarProjector
 from .taxonomy.model import TaxonomyProjector
 from .taxonomy_rels.model import TaxonomyWithRelationsProjector
 from .categorical.model import CategoricalProjector
+from .gda.model import GDAProjector
 from .factory import projector_factory