+++ mowl/projection/factory.py
@@ -3,9 +3,10 @@ from mowl.projection.taxonomy_rels.model import TaxonomyWithRelationsProjector
 from mowl.projection.dl2vec.model import DL2VecProjector
 
 from mowl.projection.owl2vec_star.model import OWL2VecStarProjector
+from mowl.projection.gda.model import GDAProjector
 
 
-PARSING_METHODS = ["taxonomy", "taxonomy_rels", "dl2vec", "owl2vecstar"]
+PARSING_METHODS = ["taxonomy", "taxonomy_rels", "dl2vec", "owl2vecstar", "gda"]
 
 
 def projector_factory(method_name, taxonomy=False, bidirectional_taxonomy=False,
@@ -22,6 +23,9 @@ def projector_factory(method_name, taxonomy=False, bidirectional_taxonomy=False,
     elif method_name == "owl2vecstar":
         return OWL2VecStarProjector(bidirectional_taxonomy=bidirectional_taxonomy,
                                     include_literals=include_literals, only_taxonomy=only_taxonomy)
+    elif method_name == "gda":
+        return GDAProjector(bidirectional_taxonomy=bidirectional_taxonomy,
+                            include_literals=include_literals, only_taxonomy=only_taxonomy)
     else:
         raise Exception(f"Graph generation method {method_name} unrecognized. Recognized methods \
 are: {PARSING_METHODS}")